+++ src/pat/calendar/calendar.js
@@ -63,6 +63,9 @@ parser.addArgument("event-sources", [], undefined, true);
 parser.addArgument("pat-inject-source", null);
 parser.addArgument("pat-inject-target", null);
 
+// pat-modal support for individual events
+parser.addArgument("pat-modal-class", null);
+
 // pat-tooltip support for individual events
 parser.addArgument("pat-tooltip-source", null, ["ajax", null]);
 
@@ -338,6 +341,17 @@ export default Base.extend({
             args.el.setAttribute(
                 "data-pat-inject",
                 `target: ${target || "body"}; source: ${source || "body"}`
+            );
+            do_scan = true;
+        }
+
+        // pat-modal support
+        if (args.event.url && this.options.pat["modal-class"]) {
+            // Only set pat-modal if event has a url
+            args.el.classList.add("pat-modal");
+            args.el.setAttribute(
+                "data-pat-modal",
+                `class: ${this.options.pat["modal-class"]}`
             );
             do_scan = true;
         }